+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- fix pagination for course collection
+
 ## v0.1.25
 
 - Add local data attribute `attendees`